src/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
transformers==4.36.2
torch
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
surya-ocr